jetpack-mu-wpcom: Fix test for WP 6.5 (#36720)
Now that WordPress 6.5 is released, WorDBless is bringing in that
version for package tests. This resulted in a failure in one of the
tests in packages/jetpack-mu-wpcom due to some extra classes on an HTML
tag.

Update the test to expect these extra classes.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8539755313