Added PHP 7.0 testing in WordPress previous. (#38356)
With the release of WordPress 6.6, which drops support for PHP 7.0 and
7.1, we need to switch our PHP 7.0 test to use WordPress "previous"
(currently 6.4, soon 6.5) instead of "latest".

This also needed code to downgrade the roots/wordpress package used with
WorDBless when running with WordPress previous, which we hadn't needed
before.

Then that picked up one test that had been updated for WordPress 6.5
back when that was released and now fails in the "previous" run.

And also the Jetpack Legacy Full Sync tests need to be bumped to run
with PHP 7.2, since 7.0 is on its way out.

---------

Co-authored-by: Brad Jorsch <brad.jorsch@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/9963129560

Upstream-Ref: Automattic/jetpack@ac89aacf6b1eff582f530a1b38cb824acbc0049b