Global Styles: Move frontend notice from launch bar to admin bar (#45417)
When a site that has limited Global Styles, we currently show an upsell in the launch banner on the front end (for admin users). However, the launch banner does not exist on Atomic sites, so we need a different place for the front end upsell when we'll extend Limited Global Styles to Atomic.

This PRs moves the upsell to the admin bar (which is where we moved the launch button too), instead of recreating the launch banner on Atomic sites.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/18404852008

Upstream-Ref: Automattic/jetpack@316659f236ee307ca1c16663aa9d1b5ac9104c19